Improve Treq req. handling:
in case backend.HeaderAPIKey api key header is present,
the Treq proxy will take a respective value and use it as
'cncAuth.sessionCookieName' session cookie. This will allow
us to make client-side requests to backends requiring special
session ID of the same name as the end-user one but with a different
value. This happens e.g. in WaG where users have their session ID
but the session ID cannot be used to access backends. For that,
we have a special 'wag' user session which is handled internally.